==> legacy/analysis/scripts/sc-processing/1_cell_QC.Rmd <==
---
title: "cycle 1 - cell QC"
output: 
  html_notebook:
    toc: true
    toc_float: true
    code_folding: show
    theme: united
---

```{r setup}
library(tidyverse)
library(openxlsx)

# SC methods
library(scran)
library(scater)
library(bluster)

# Doublet scoring
library(scDblFinder)

# Annotations
library(SingleR)

# Variable with random seed value used throughout the analysis for reproducibility
random_seed_value = 17

# Parallelization
library(BiocParallel)
bpp <- MulticoreParam()

# Path to project folder
wd <- here::here()

iteration_input = "0_cells"
sce.dir <- file.path(wd, "scRNAseq/analysis/outputs/", iteration_input)
```

```{r}
patients <- c("P009", "P022", "P027", "P087", "P069")
timepoints <- c("DG", "REL")
compartments <- c("PBMC", "BM", "LN", "TON", "GUT", "ASC")

samples.df <- read.xlsx(paste(sce.dir, "cell_metadata.xlsx", sep = "/")) %>% 
  mutate(
    patient = factor(patient, levels = patients),
    timepoint = factor(timepoint, levels = timepoints),
    compartment = factor(compartment, levels = compartments)
  )
samples.df
```

```{r}
joined_sce <- readRDS(file = paste(sce.dir, "sce.rds", sep = "/"))
joined_sce$Sample <- factor(joined_sce$Sample, levels = samples.df$sample)
joined_sce$patient <- factor(joined_sce$patient, levels = patients)
joined_sce$timepoint <- factor(joined_sce$timepoint, levels = timepoints)
joined_sce$compartment <- factor(joined_sce$compartment, levels = compartments)
joined_sce$repl <- word(joined_sce$Sample, 4, sep = "_")

joined_sce
```

# Processing

## Rownames conversion

```{r}
rownames(joined_sce) <- rowData(joined_sce)[["Symbol"]]
joined_sce
```

## Normalization

```{r warning=FALSE}
set.seed(random_seed_value)
clusters_igraph <- quickCluster(joined_sce,
                                method = "igraph",
                                BPPARAM = bpp)
joined_sce$clusters_igraph <- clusters_igraph

set.seed(random_seed_value)
joined_sce <- computeSumFactors(joined_sce,
                                cluster = clusters_igraph,
                                min.mean = 0.1,
                                BPPARAM = bpp)
joined_sce <- logNormCounts(joined_sce,
                            size.factors = joined_sce$sizeFactor,
                            BPPARAM = bpp)
joined_sce
```

## HVG detection

```{r warning=FALSE}
# Feature selection
set.seed(random_seed_value)
variance_decomposition <- modelGeneVar(joined_sce, BPPARAM = bpp)
top_HVG <- getTopHVGs(variance_decomposition, fdr.threshold = 0.25)
```

```{r fig.height=5, fig.width=8, warning=FALSE}
variance_df <- as.data.frame(variance_decomposition) %>%
  rownames_to_column("Symbol") %>%
  left_join(., as.data.frame(rowData(joined_sce)), by = "Symbol") %>%
  mutate(HVG = case_when(Symbol %in% top_HVG ~ TRUE, TRUE ~ FALSE))

ggplot(variance_df,
       aes(x = mean, y = bio, color = HVG)) +
  scale_color_manual(values = c("black", "red")) +
  ggrepel::geom_text_repel(
    data = dplyr::filter(variance_df, bio > 0.75),
    aes(label = Symbol), min.segment.length = 0,
    color = "gray30"
  ) +
  geom_point(shape = 1) +
  cowplot::theme_cowplot() +
  labs(
    title = "Variance decomposition plot",
    x = "Mean of log-expression",
    y = "Variance of log-expression",
    color = "Biological\nvariance"
  )
```

# Dimensionality reduction

## PCA

```{r}
# PCA
set.seed(random_seed_value)
joined_sce <- denoisePCA(joined_sce,
                         subset.row = top_HVG,
                         technical = variance_decomposition)
percent.var <- attr(reducedDim(joined_sce), "percentVar")

# chose a number of PCA dimensions which explain more than 0.5% variance
chosen.PC.num <-
  percent.var[which(percent.var > 0.5)] %>% length()

plot(percent.var, log="y", xlab="PC", ylab="Variance explained (%)")
abline(v = chosen.PC.num, col = "blue")
```

```{r, fig.width=7, fig.height=6}
plotPCA(joined_sce, color_by = "Sample")
```


```{r, fig.width=8.5, fig.height=8}
ggpubr::ggarrange(
  plotPCA(joined_sce, color_by = "discard"),
  plotPCA(joined_sce, color_by = "scDblFinder.calls"),
  plotPCA(joined_sce, color_by = "MS4A1"),
  plotPCA(joined_sce, color_by = "CCND1")
)
```

## UMAP

```{r}
set.seed(random_seed_value)
joined_sce <- runUMAP(
  joined_sce,
  pca = chosen.PC.num,
  n_neighbors = 30,
  min_dist = 0.2,
  dimred = "PCA"
)
```

```{r, fig.width=6.5, fig.height=5}
plotUMAP(joined_sce, color_by = "Sample")
```

```{r, fig.width=13.5, fig.height=5.5}
plotUMAP(joined_sce,
         color_by = "compartment",
         other_fields = c("patient", "timepoint")
         ) +
  facet_grid(timepoint ~ patient) +
  theme_bw()
```

```{r, fig.width=6, fig.height=5}
plotUMAP(joined_sce, color_by = "detected")
```

```{r, fig.width=6, fig.height=5}
plotUMAP(joined_sce, color_by = "discard") +
  annotate(
    "rect",
    xmin = -5,
    xmax = 0,
    ymin = 2,
    ymax = 10,
    color = "red",
    fill = NA
  )
```

```{r fig.height=5, fig.width=6, message=FALSE}
plotUMAP(joined_sce, color_by = "subsets_Mito_percent") +
  scale_color_viridis_c(option = "B") +
  labs(color = "MT- genes %")
```


```{r, fig.width=6, fig.height=5}
plotUMAP(joined_sce, color_by = "scDblFinder.calls") +
  annotate(
    "rect",
    xmin = -7,
    xmax = -3,
    ymin = 1,
    ymax = -5.5,
    color = "red",
    fill = NA
  )
```

```{r, fig.width=11, fig.height=6}
plotUMAP(joined_sce,
         color_by = "discard",
         other_fields = "Sample") +
  facet_wrap( ~ Sample, ncol = 6) +
  annotate(
    "rect",
    xmin = -5,
    xmax = 0,
    ymin = 2,
    ymax = 10,
    color = "red",
    fill = NA
  )
```

```{r, fig.width=11, fig.height=6}
plotUMAP(joined_sce,
         color_by = "scDblFinder.calls",
         other_fields = "Sample") +
  facet_wrap(~ Sample, ncol = 6) +
  annotate(
    "rect",
    xmin = -7,
    xmax = -3,
    ymin = 1,
    ymax = -5.5,
    color = "red",
    fill = NA
  )
```

# Annotation

## Unsupervized clustering

```{r, fig.width=6, fig.height=5}
KNN = 30

set.seed(random_seed_value)
nn.clust <- clusterCells(
  joined_sce,
  use.dimred = "PCA",
  full = TRUE,
  BLUSPARAM = bluster::SNNGraphParam(
    k = KNN,
    type = "rank",
    cluster.fun = "louvain",
    BPPARAM = bpp
  )
)

joined_sce[[paste0("clusters_louvain_K", KNN)]] <- nn.clust$clusters

plotUMAP(joined_sce,
         color_by = paste0("clusters_louvain_K", KNN),
         text_by = paste0("clusters_louvain_K", KNN))
```

```{r, fig.width=6, fig.height=5}
KNN = 40

set.seed(random_seed_value)
nn.clust <- clusterCells(
  joined_sce,
  use.dimred = "PCA",
  full = TRUE,
  BLUSPARAM = bluster::SNNGraphParam(
    k = KNN,
    type = "rank",
    cluster.fun = "louvain",
    BPPARAM = bpp
  )
)

joined_sce[[paste0("clusters_louvain_K", KNN)]] <- nn.clust$clusters

plotUMAP(joined_sce,
         color_by = paste0("clusters_louvain_K", KNN),
         text_by = paste0("clusters_louvain_K", KNN)
         )
```

```{r, fig.width=6, fig.height=5}
KNN = 45

set.seed(random_seed_value)
nn.clust <- clusterCells(
  joined_sce,
  use.dimred = "PCA",
  full = TRUE,
  BLUSPARAM = bluster::SNNGraphParam(
    k = KNN,
    type = "rank",
    cluster.fun = "louvain",
    BPPARAM = bpp
  )
)

joined_sce[[paste0("clusters_louvain_K", KNN)]] <- nn.clust$clusters

plotUMAP(joined_sce,
         color_by = paste0("clusters_louvain_K", KNN),
         text_by = paste0("clusters_louvain_K", KNN)
         )
```

```{r, fig.width=6, fig.height=5}
KNN = 50

set.seed(random_seed_value)
nn.clust <- clusterCells(
  joined_sce,
  use.dimred = "PCA",
  full = TRUE,
  BLUSPARAM = bluster::SNNGraphParam(
    k = KNN,
    type = "rank",
    cluster.fun = "louvain",
    BPPARAM = bpp
  )
)

joined_sce[[paste0("clusters_louvain_K", KNN)]] <- nn.clust$clusters

plotUMAP(joined_sce,
         color_by = paste0("clusters_louvain_K", KNN),
         text_by = paste0("clusters_louvain_K", KNN)
         )
```

```{r, fig.width=6, fig.height=5}
joined_sce$remove <- colData(joined_sce) %>% 
  as.data.frame() %>% 
  mutate(
    remove = case_when(
      discard == TRUE | scDblFinder.calls == "doublet" |
        clusters_louvain_K40 %in% c(4, 14, 16) ~ TRUE,
      TRUE ~ FALSE
    )
  ) %>% pull(remove)

plotUMAP(joined_sce,
         color_by = "remove"
         )
```

# Filter undamaged cells and save

```{r}
sce.save <- joined_sce[, !(joined_sce$remove)]
sce.save$remove <- NULL
sce.save$scDblFinder.calls <- NULL
sce.save$discard <- NULL
sce.save$low_lib_size <- NULL
sce.save$low_n_features <- NULL
sce.save$high_subsets_Mito_percent <- NULL
sce.save$clusters_igraph <- NULL
```

Counting

```{r}
samples.df <- colData(sce.save) %>% 
  as.data.frame() %>% 
  remove_rownames() %>% 
  select(Sample) %>% 
  group_by(Sample) %>% 
  summarise(N_cells_QCfiltered = n()) %>% 
  left_join(
    samples.df, ., by = c("sample" = "Sample")
  )
samples.df
```

```{r}
iteration_output = "1_cells_QC"
storage.dir <- file.path(wd, "scRNAseq/analysis/outputs/", iteration_output)

dir.create(storage.dir, showWarnings = F, recursive = T)
saveRDS(sce.save,
        file = file.path(storage.dir, "sce.rds")
        )
write.xlsx(samples.df,
           file.path(storage.dir, "cell_metadata.xlsx")
           )
```

```{r}
sessionInfo()
```
