==> legacy/analysis/scripts/sc-processing/2_cell_types.Rmd <==
---
title: "cycle 2 - cell type annotation"
output: 
  html_notebook:
    toc: true
    toc_float: true
    code_folding: show
    theme: united
---

```{r setup}
library(tidyverse)
library(openxlsx)

# SC methods
library(scran)
library(scater)
library(bluster)

# Annotations
library(SingleR)
library(Seurat)
library(tricycle)

# Variable with random seed value used throughout the analysis for reproducibility
random_seed_value = 17

# Parallelization
library(BiocParallel)
bpp <- MulticoreParam(workers = 60)

# Path to project folder
wd <- here::here()
source(file.path(wd, "scRNAseq/analysis/scripts", "colors.R"))

iteration_input = "1_cells_QC"
sce.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_input)
```

```{r}
patients <- c("P009", "P022", "P027", "P087", "P069")
timepoints <- c("DG", "REL")
compartments <- c("PBMC", "BM", "LN", "TON", "GUT", "ASC")

df.samples <- read.xlsx(file.path(sce.dir, "cell_metadata.xlsx")) %>% 
  mutate(
    patient = factor(patient, levels = patients),
    timepoint = factor(timepoint, levels = timepoints),
    compartment = factor(compartment, levels = compartments)
  )
df.samples
```

```{r}
sce <- readRDS(file = file.path(sce.dir, "sce.rds"))
sce
```

# Processing

## Normalization

```{r warning=FALSE}
set.seed(random_seed_value)
clusters_igraph <- quickCluster(sce,
                                method = "igraph",
                                BPPARAM = bpp)
sce$clusters_igraph <- clusters_igraph

set.seed(random_seed_value)
sce <- computeSumFactors(sce,
                         cluster = clusters_igraph,
                         min.mean = 0.1,
                         BPPARAM = bpp)
sce <- logNormCounts(sce,
                     size.factors = sce$sizeFactor,
                     BPPARAM = bpp)
sce
```

## HVG detection

```{r warning=FALSE}
set.seed(random_seed_value)
variance_decomposition <- modelGeneVar(sce, BPPARAM = bpp)
top_HVG <- getTopHVGs(variance_decomposition, fdr.threshold = 0.25)
```

```{r fig.height=6, fig.width=10, warning=FALSE}
# removing IGLC2 and IGKC as their variation is given by the patient of origin and not cell biology
top_HVG <- top_HVG[!top_HVG %in% c("IGKC")]

variance_df <- as.data.frame(variance_decomposition) %>%
  rownames_to_column("Symbol") %>%
  left_join(., as.data.frame(rowData(sce)), by = "Symbol") %>%
  mutate(HVG = case_when(Symbol %in% top_HVG ~ TRUE, TRUE ~ FALSE))

ggplot(variance_df,
       aes(x = mean, y = bio, color = HVG)) +
  scale_color_manual(values = c("black", "red")) +
  ggrepel::geom_text_repel(
    data = dplyr::filter(variance_df, bio > 0.75),
    aes(label = Symbol), min.segment.length = 0,
    color = "gray30"
  ) +
  geom_point(shape = 1) +
  cowplot::theme_cowplot() +
  labs(
    title = "Variance decomposition plot",
    x = "Mean of log-expression",
    y = "Variance of log-expression",
    color = "Biological\nvariance"
  )
```

# Dimensionality reduction

## PCA

```{r}
# PCA
set.seed(random_seed_value)
sce <- denoisePCA(sce,
                         subset.row = top_HVG,
                         technical = variance_decomposition)
percent.var <- attr(reducedDim(sce), "percentVar")

# chose a number of PCA dimensions which explain more than 0.5% variance
chosen.PC.num <-
  percent.var[which(percent.var > 0.25)] %>% length()

plot(percent.var, log="y", xlab="PC", ylab="Variance explained (%)")
abline(v = chosen.PC.num, col = "blue")
```

```{r, fig.width=6.5, fig.height=5}
plotPCA(sce, color_by = "Sample")
```

```{r, fig.width=6, fig.height=5}
plotPCA(sce, color_by = "sum")
```

```{r, fig.width=6, fig.height=5}
plotPCA(sce, color_by = "detected")
```

```{r, fig.width=6, fig.height=5}
plotPCA(sce, color_by = "MS4A1")
```

```{r, fig.width=6, fig.height=5}
plotPCA(sce, color_by = "CD3G")
```

```{r, fig.width=6, fig.height=5}
plotPCA(sce, color_by = "CCND1")
```

## tSNE

```{r, fig.width=6.5, fig.height=5}
set.seed(random_seed_value)
sce <-
  runTSNE(
    sce,
    BPPARAM = bpp
  )
plotReducedDim(sce,
               dimred = "TSNE",
               color_by = "Sample")
```

```{r, fig.width=13, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE",
               color_by = "compartment",
               other_fields = c("patient", "timepoint")
               ) +
  facet_grid(timepoint ~ patient) +
  theme_bw()
```

```{r, fig.width=6, fig.height=5}
plotTSNE(sce, color_by = "MS4A1")
```

```{r, fig.width=6, fig.height=5}
plotTSNE(sce, color_by = "CCND1")
```

# Annotations

## Cell types

```{r warning=FALSE, fig.width=6, fig.height=5}
ref <- celldex::MonacoImmuneData()

set.seed(random_seed_value)
predicted.types.main <- SingleR(
  test = sce,
  ref = ref,
  labels = ref$label.main,
  de.method = "classic"
)

# assign cell types to sce
sce$cell_type_main <- predicted.types.main$labels

plotTSNE(sce,
         color_by = "cell_type_main")
```

```{r fig.width=6, fig.height=5}
plotTSNE(sce,
         color_by = "CCND1")
```

```{r, fig.width=13, fig.height=4}
plotTSNE(sce,
         color_by = "cell_type_main",
         other_fields = c("patient", "timepoint")
         ) +
  theme_bw() +
  facet_grid(timepoint ~ patient)
```

```{r warning=FALSE, fig.width=8, fig.height=5}
ref <- celldex::MonacoImmuneData()

set.seed(random_seed_value)
predicted.types.fine <- SingleR(
  test = sce,
  ref = ref,
  labels = ref$label.fine,
  de.method = "classic"
)

# assign cell types to sce
sce$cell_type_fine <- predicted.types.fine$labels

plotTSNE(sce,
         color_by = "cell_type_fine")
```

## Cell cycle phase

```{r, fig.width=6, fig.height=5}
sce.to.seu <- sce

s.genes <- as.data.frame(rowData(sce.to.seu)) %>%
  filter(Symbol %in% cc.genes.updated.2019$s.genes) %>%
  pull(ID)

g2m.genes <- as.data.frame(rowData(sce.to.seu)) %>%
  filter(Symbol %in% cc.genes.updated.2019$g2m.genes) %>%
  pull(ID)

rownames(sce.to.seu) <- rowData(sce.to.seu)[["ID"]]
seu <-
  SeuratObject::as.Seurat(sce.to.seu, counts = "counts", data = "logcounts")

seu.cc.scores <- CellCycleScoring(
  seu,
  s.features = s.genes,
  g2m.features = g2m.genes,
  set.ident = TRUE
)

sce$cc.phase.seurat <- left_join(
  data.frame("barcode" = colnames(sce)),
  data.frame("cc.phase" = seu.cc.scores$Phase) %>%
    rownames_to_column("barcode")
) %>% pull(cc.phase) %>%
  factor(., levels = c("G1", "S", "G2M"))

plotTSNE(sce, color_by = "cc.phase.seurat")
```

```{r, fig.width=6, fig.height=5}
sce <- project_cycle_space(sce,
                           species = "human",
                           gname.type = "SYMBOL")
plotReducedDim(sce, dimred = "tricycleEmbedding",
               color_by = "cc.phase.seurat")
```

```{r, fig.width=7, fig.height=5}
center_PCs <- c(-3.5,-0.2)

sce <- estimate_cycle_position(sce,
                               center.pc1 = center_PCs[1],
                               center.pc2 = center_PCs[2])

p <- plot_emb_circle_scale(sce,
                           dimred = "tricycleEmbedding",
                           point.size = 3.5,
                           point.alpha = 0.9) +
  annotate(
    "point",
    x = center_PCs[1],
    y = center_PCs[2],
    size = 4,
    color = "red",
    shape = 8,
    stroke = 1
  ) +
  labs(title = NULL) +
  theme_bw(base_size = 14)
legend <- circle_scale_legend(text.size = 4.5, alpha = 0.9)
cowplot::plot_grid(p, legend, ncol = 2, rel_widths = c(1, 0.4))
```

```{r, fig.width=6, fig.height=5}
phase.definitions = c(
  "G1" = 0.2 * pi,
  "S" = 0.72 * pi,
  "G2" = 0.975 * pi,
  "G2.M" = 1.12 * pi,
  "M.G1" = 1.5 * pi,
  "G0" = 1.87 * pi
)

TCP.df <- data.frame(colname = colnames(sce),
                     TCP = sce$tricyclePosition,
                     Phase = NA
                     )

for(position in names(phase.definitions)){
  TCP.df[TCP.df$TCP > phase.definitions[[position]] ,]$Phase <- position
}

TCP.df[is.na(TCP.df$Phase), ]$Phase <- "G0"

TCP.df$Phase <- factor(TCP.df$Phase,
                       levels = c("G0", "G1", "S", "G2", "G2.M", "M.G1"))

sce$tricyclePhase <- TCP.df$Phase

ggcells(sce,
        mapping = aes(x = tricycleEmbedding.1,
                      y = tricycleEmbedding.2
                      )
        ) +
  geom_point(
    aes(color = tricyclePhase)
  ) +
  geom_density2d(
    data = as.data.frame(reducedDims(sce)[["tricycleEmbedding"]]),
    aes(x = PC1, y = PC2), adjust = 2,
    color = "black"
  ) +
  annotate("point",
           x = center_PCs[1],
           y = center_PCs[2],
           size = 4,
           color = "black", shape = 4,
           stroke = 1
           )
```

```{r, fig.width=6, fig.height=5}
plotTSNE(sce, color_by = "tricyclePhase")
```

# Batch correction

## For sample and CC phase

```{r warning=FALSE}
sce$correct_for <- paste(sce$Sample,
                         sce$tricyclePhase, sep = "_")

set.seed(random_seed_value)
sce <- batchelor::correctExperiments(sce,
                                     batch = sce$correct_for,
                                     PARAM = batchelor::FastMnnParam(BPPARAM = bpp))

set.seed(random_seed_value)
sce <- runTSNE(sce,
               dimred = "corrected",
               name = "TSNE_corrected_SamplePhase",
               BPPARAM = bpp)

set.seed(random_seed_value)
sce <-
  runUMAP(
    sce,
    n_neighbors = 15,
    min_dist = 0.2,
    dimred = "corrected",
    name = "UMAP_corrected_SamplePhase",
    BPPARAM = bpp
  )
```

```{r, fig.width=6, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "Sample")
```

```{r, fig.width=6, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "cell_type_main")
```

```{r, fig.width=6, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "CCND1")
```

```{r, fig.width=6, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "tricyclePhase")
```

## Unsupervized clustering

```{r}
sce$clusters_louvain_K30 <- NULL
sce$clusters_louvain_K40 <- NULL
sce$clusters_louvain_K45 <- NULL
sce$clusters_louvain_K50 <- NULL
sce$clusters_louvain_K100 <- NULL
```

```{r, fig.width=6, fig.height=5}
KNN = 50

set.seed(random_seed_value)
nn.clust <- clusterCells(
  sce,
  use.dimred = "PCA",
  full = TRUE,
  BLUSPARAM = bluster::SNNGraphParam(
    k = KNN,
    type = "rank",
    cluster.fun = "louvain",
    BPPARAM = bpp
  )
)

sce[[paste0("clusters_louvain_K", KNN)]] <- nn.clust$clusters

plotTSNE(sce,
         color_by = paste0("clusters_louvain_K", KNN),
         text_by = paste0("clusters_louvain_K", KNN)
         )
```

```{r}
markers <- scoreMarkers(sce, lfc = 2,
                        groups = sce$clusters_louvain_K50,
                        block = sce$Sample,
                        BPPARAM = bpp
                        )
```

```{r, fig.width=7, fig.height=7}
clusters <- levels(sce$clusters_louvain_K50)

top.genes <- markers %>% 
  lapply(., as.data.frame) %>% 
  lapply(., rownames_to_column, "genesymbol") %>% 
  lapply(., arrange, desc(min.logFC.cohen)) %>% 
  bind_rows(.id = "cluster") %>% 
  mutate(cluster = factor(cluster, levels = clusters)) %>% 
  group_by(cluster) %>% 
  slice_head(n = 5)

N.clusters <- length(clusters)
cluster.colors <- get_palettes("tableau20")[1:N.clusters]
names(cluster.colors) <- clusters
  
annotation_colors <- list(cluster = cluster.colors)
annotation_clusters <- data.frame(cluster = factor(names(cluster.colors),
                                                   levels = names(cluster.colors)))
annotation_genes <- top.genes[!duplicated(top.genes$genesymbol), ] %>%
  dplyr::select(cluster, genesymbol) %>%
  column_to_rownames("genesymbol") %>%
  mutate(cluster = factor(cluster, levels = names(cluster.colors)))

scater::plotGroupedHeatmap(
    sce, 
    #main = "title",
    features = unique(top.genes$genesymbol),
    group = "clusters_louvain_K50",
    center = TRUE,
    cluster_rows = FALSE,
    cluster_cols = FALSE,
    angle_col = 0,
    annotation_colors = annotation_colors,
    annotation_row = annotation_genes
  )
```

# Cell labelling

```{r, fig.width=10, fig.height=4}
tSNE.clusters.types <- ggpubr::ggarrange(
  plotTSNE(
    sce,
    color_by = paste0("clusters_louvain_K", KNN),
    text_by = paste0("clusters_louvain_K", KNN)
  ),
  plotReducedDim(sce, dimred = "TSNE", color_by = "cell_type_main"),
  labels = c("A", "B"),
  nrow = 1
)
tSNE.clusters.types
```

```{r}
sce$cell_type_manual <- "B cell"

all_clusters <- seq(1:length(levels(sce$clusters_louvain_K50)))

# specify non-B cell clusters
T_cells_clusters = c(1, 4, 5)
Monocytes_clusters = c(3)

colData(sce) %>%
  as.data.frame() %>%
  select(clusters_louvain_K50, cell_type_manual) %>%
  mutate(
    cell_type_manual = case_when(
      clusters_louvain_K50 %in% T_cells_clusters ~ "T cell",
      clusters_louvain_K50 %in% Monocytes_clusters ~ "Monocyte",
      
      TRUE ~ cell_type_manual
    )
  ) %>%
  pull(cell_type_manual) -> sce$cell_type_manual

sce$cell_type_manual <- factor(sce$cell_type_manual, levels = c("B cell", "T cell", "Monocyte"))

tSNE.labels <- plotReducedDim(sce,
               dimred = "TSNE",
               color_by = "cell_type_manual",
               other_fields = "patient") +
  theme_bw() +
  facet_wrap(~ patient)
tSNE.labels
```

```{r, fig.width=9, fig.height=8}
tSNE.comp <- ggpubr::ggarrange(
  tSNE.clusters.types,
  tSNE.labels,
  labels = c("", "C"),
  ncol = 1, heights = c(2, 3)
)
tSNE.comp
```

```{r}
figures.out.dir <- file.path(wd, "scRNAseq/figures")

tiff(
  filename = file.path(figures.out.dir, "SFig2.Cell_type_annotation.tiff"),
  units = "in",
  width = 9,
  height = 8,
  res = 300,
  bg = "white"
)
tSNE.comp
dev.off()
```

```{r}
markers <- scoreMarkers(sce, lfc = 2,
                        groups = sce$cell_type_manual,
                        block = sce$Sample,
                        BPPARAM = bpp
                        )
```

```{r, fig.width=5, fig.height=6}
clusters <- levels(sce$cell_type_manual)

top.genes <- markers %>% 
  lapply(., as.data.frame) %>% 
  lapply(., rownames_to_column, "genesymbol") %>% 
  lapply(., arrange, desc(min.logFC.cohen)) %>% 
  bind_rows(.id = "cluster") %>% 
  mutate(cluster = factor(cluster, levels = clusters)) %>% 
  group_by(cluster) %>% 
  slice_head(n = 10)

N.clusters <- length(clusters)
cluster.colors <- get_palettes("tableau20")[1:N.clusters]
names(cluster.colors) <- clusters
  
annotation_colors <- list("cell type" = cluster.colors)
annotation_clusters <- data.frame(cluster = factor(names(cluster.colors),
                                                   levels = names(cluster.colors)))
annotation_genes <- top.genes[!duplicated(top.genes$genesymbol), ] %>%
  dplyr::select("cell type" = cluster, genesymbol) %>%
  column_to_rownames("genesymbol") %>%
  mutate(`cell type` = factor(`cell type`, levels = names(cluster.colors)))

heatmap.cell_type <- scater::plotGroupedHeatmap(
    sce, 
    #main = "title",
    features = unique(top.genes$genesymbol),
    group = "cell_type_manual",
    center = TRUE,
    cluster_rows = FALSE,
    cluster_cols = FALSE,
    angle_col = 0,
    annotation_colors = annotation_colors,
    annotation_row = annotation_genes
  )
heatmap.cell_type
```

```{r}
tiff(
  filename = file.path(figures.out.dir, "SFig2.heatmap.tiff"),
  units = "in",
  width = 5,
  height = 6,
  res = 300,
  bg = "white"
)
grid::grid.newpage()
grid::grid.draw(heatmap.cell_type)
dev.off()
```


```{r fig.width=5, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE",
               color_by = "cell_type_manual",
               text_by = "cell_type_manual"
               ) +
  theme(legend.position = "none")
```

# Save

```{r}
iteration_output = "2_cell_types"
storage.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_output)
dir.create(storage.dir, showWarnings = F, recursive = T)
saveRDS(sce,
        file = file.path(storage.dir, "sce.rds")
        )
write.xlsx(df.samples,
           file.path(storage.dir, "cell_metadata.xlsx")
           )
```

# Session info

```{r}
sessionInfo()
```
