==> legacy/analysis/scripts/sc-processing/3_Bcells.Rmd <==
---
title: "cycle 3 - B cell type annotation"
output: 
  html_notebook:
    toc: true
    toc_float: true
    code_folding: show
    theme: united
---

```{r setup}
library(tidyverse)
library(openxlsx)
library(patchwork)

# SC methods
library(scran)
library(scater)
library(bluster)

# Annotations
library(SingleR)
library(Seurat)
library(tricycle)

# Variable with random seed value used throughout the analysis for reproducibility
random_seed_value = 17

# Parallelization
library(BiocParallel)
bpp <- MulticoreParam(workers = 60)

# Path to project folder
wd <- here::here()

iteration_input = "2_cell_types"
sce.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_input)
```

```{r}
patients <- c("P009", "P022", "P027", "P087", "P069")
timepoints <- c("DG", "REL")
compartments <- c("PBMC", "BM", "LN", "TON", "GUT", "ASC")

samples.df <- read.xlsx(file.path(sce.dir, "cell_metadata.xlsx")) %>% 
  mutate(
    patient = factor(patient, levels = patients),
    timepoint = factor(timepoint, levels = timepoints),
    compartment = factor(compartment, levels = compartments)
  )
samples.df
```

```{r}
sce <- readRDS(file = file.path(sce.dir, "sce.rds"))
sce
```

# Filtering

```{r, fig.width=6, fig.height=5}
plotReducedDim(sce,
               dimred = "TSNE",
               color_by = "cell_type_manual",
               text_by = "cell_type_manual"
               )
```

## Apply filters

```{r}
sce.filtered <- sce[, sce$cell_type_manual == "B cell"]
sce.filtered
```

# Processing

## Normalization

```{r warning=FALSE}
set.seed(random_seed_value)
clusters_igraph <- quickCluster(sce.filtered,
                                method = "igraph",
                                BPPARAM = bpp)
sce.filtered$clusters_igraph <- clusters_igraph

set.seed(random_seed_value)
sce.filtered <- computeSumFactors(sce.filtered,
                                cluster = clusters_igraph,
                                min.mean = 0.1,
                                BPPARAM = bpp)
sce.filtered <- logNormCounts(sce.filtered,
                            size.factors = sce.filtered$sizeFactor,
                            BPPARAM = bpp)
sce.filtered
```

## HVG detection

```{r warning=FALSE}
set.seed(random_seed_value)
variance_decomposition <- modelGeneVar(sce.filtered, BPPARAM = bpp)
top_HVG <- getTopHVGs(variance_decomposition, fdr.threshold = 0.25)
```

```{r fig.height=6, fig.width=10, warning=FALSE}
# removing IGLC2 and IGKC as their variation is given by the patient of origin and not cell biology
top_HVG <- top_HVG[!top_HVG %in% c("IGKC", "IGLC2")]

variance_df <- as.data.frame(variance_decomposition) %>%
  rownames_to_column("Symbol") %>%
  left_join(., as.data.frame(rowData(sce.filtered)), by = "Symbol") %>%
  mutate(HVG = case_when(Symbol %in% top_HVG ~ TRUE, TRUE ~ FALSE))

ggplot(variance_df,
       aes(x = mean, y = bio, color = HVG)) +
  scale_color_manual(values = c("black", "red")) +
  ggrepel::geom_text_repel(
    data = dplyr::filter(variance_df, bio > 0.75),
    aes(label = Symbol), min.segment.length = 0,
    color = "gray30"
  ) +
  geom_point(shape = 1) +
  cowplot::theme_cowplot() +
  labs(
    title = "Variance decomposition plot",
    x = "Mean of log-expression",
    y = "Variance of log-expression",
    color = "Biological\nvariance"
  )
```

# Dimensionality reduction

## PCA

```{r}
# PCA
set.seed(random_seed_value)
sce.filtered <- denoisePCA(sce.filtered,
                           subset.row = top_HVG,
                           technical = variance_decomposition)
percent.var <- attr(reducedDims(sce.filtered)[["PCA"]], "percentVar")

# chose a number of PCA dimensions which explain more than 0.5% variance
chosen.PC.num <-
  percent.var[which(percent.var > 0.25)] %>% length()

plot(percent.var,
     log = "y",
     xlab = "PC",
     ylab = "Variance explained (%)")
abline(v = chosen.PC.num, col = "blue")
```

```{r, fig.width=12, fig.height=6}
(
  plotPCA(sce.filtered, color_by = "Sample") +
    plotPCA(sce.filtered, color_by = "sum") +
    plotPCA(sce.filtered, color_by = "detected")
) /
  (plotPCA(sce.filtered, color_by = "MS4A1") +
     plotPCA(sce.filtered, color_by = "CCND1") +
     plotPCA(sce.filtered, color_by = "cell_type_fine"))
```

## tSNE

```{r, fig.width=6.5, fig.height=5}
set.seed(random_seed_value)
sce.filtered <-
  runTSNE(
    sce.filtered,
    BPPARAM = bpp
  )
plotReducedDim(sce.filtered,
               dimred = "TSNE",
               color_by = "Sample")
```

```{r, fig.width=10, fig.height=4}
plotReducedDim(sce.filtered,
               dimred = "TSNE",
               color_by = "compartment",
               other_fields = c("patient", "timepoint")
               ) +
  facet_grid(timepoint ~ patient) +
  theme_bw()
```

```{r, fig.width=6, fig.height=5}
plotTSNE(sce.filtered, color_by = "CCND1")
```



# Annotations

## Unsupervized clustering

```{r}
sce.filtered$clusters_louvain_K30 <- NULL
sce.filtered$clusters_louvain_K45 <- NULL
sce.filtered$clusters_louvain_K50 <- NULL
```

```{r, fig.width=6, fig.height=5}
KNN = 60

set.seed(random_seed_value)
nn.clust <- clusterCells(
  sce.filtered,
  use.dimred = "PCA",
  full = TRUE,
  BLUSPARAM = bluster::SNNGraphParam(
    k = KNN,
    type = "rank",
    cluster.fun = "louvain",
    BPPARAM = bpp
  )
)

sce.filtered[[paste0("clusters_louvain_K", KNN)]] <- nn.clust$clusters

print(plotReducedDim(
  sce.filtered,
  dimred = "TSNE",
  color_by = paste0("clusters_louvain_K", KNN),
  text_by = paste0("clusters_louvain_K", KNN)
))
```

```{r, fig.width=8, fig.height=4}
cluster_resolution = "clusters_louvain_K60"

plotTSNE(sce.filtered,
         color_by = cluster_resolution,
         text_by = cluster_resolution) + theme(legend.position = "none") +
  plotTSNE(sce.filtered, color_by = "CCND1")
```





# Marker genes expression

```{r}
str_subset(rownames(sce.filtered), "^IGLC[0-9]$")
```


```{r, fig.width=8, fig.height=6}
plotExpression(
  sce.filtered,
  x = cluster_resolution,
  color_by = cluster_resolution,
  features = c("CCND1", "IGKC", "IGLC1", "IGLC2"),
  ncol = 1
) + theme(legend.position = "none")
```

```{r, fig.width=8, fig.height=4}
plotExpression(
  sce.filtered,
  x = cluster_resolution,
  color_by = cluster_resolution,
  features = c("IGHM", "IGHD"),
  ncol = 1
) + theme(legend.position = "none")
```

```{r, fig.width=8, fig.height=5}
plotExpression(
  sce.filtered,
  x = cluster_resolution,
  color_by = cluster_resolution,
  features = c("MME", "CD24", "CD27", "CD38"),
  ncol = 1
) + theme(legend.position = "none")
```

# Cell labelling

```{r, fig.width=12, fig.height=5}
plotReducedDim(sce.filtered,
         dimred = "TSNE",
         color_by = "Sample"
         ) +

plotReducedDim(sce.filtered,
         dimred = "TSNE",
         color_by = cluster_resolution,
         text_by = cluster_resolution
         )
```

```{r}
all_clusters <- seq(1:length(levels(sce.filtered[[cluster_resolution]])))
MCLcell_clusters <- all_clusters[!all_clusters %in% c(2, 3) ]

sce.filtered$cell_type_manual <- colData(sce.filtered) %>% 
  as.data.frame() %>% 
  select(cluster = contains(cluster_resolution), cell_type_manual) %>% 
  mutate(
    cell_type_manual = case_when(
      cluster %in% MCLcell_clusters ~ "MCL",
      TRUE ~ cell_type_manual
    )
  ) %>% 
  pull(cell_type_manual)

plotTSNE(sce.filtered,
         color_by = "cell_type_manual",
         other_fields = "patient"
         ) + facet_wrap(~ patient)
```

```{r, fig.width=10, fig.height=8}
plotTSNE(sce.filtered,
         color_by = "cell_type_manual",
         other_fields = c("patient", "compartment")) +
  facet_wrap(patient ~ compartment) +
  theme_bw()
```

```{r, fig.width=10, fig.height=8}
plotTSNE(sce.filtered,
         color_by = "CCND1",
         other_fields = c("patient", "compartment")) +
  facet_wrap(patient ~ compartment) +
  theme_bw()
```

```{r}
sce.filtered[, sce.filtered$patient == "P009"] %>% 
  plotExpression(., "CCND1", x = "cell_type_manual",
                 color_by = "cell_type_manual")
```

```{r}
sce.P009 <- sce.filtered[, sce.filtered$patient == "P009"]

markers.P009 <- sce.P009 %>% 
  scoreMarkers(., lfc = 2,
               groups = sce.P009$cell_type_manual)
markers.P009[[2]] %>% 
  as.data.frame() %>% 
  rownames_to_column("genesymbol") %>% 
  filter(genesymbol == "CCND1")
```

```{r}
p009.Wilcox <- pairwiseWilcox(sce.P009, groups = sce.P009$cell_type_manual)

p009.Wilcox$statistics[[2]] %>% 
  as.data.frame() %>% 
  rownames_to_column("genesymbol") %>% 
  filter(genesymbol == "CCND1")
```

```{r}
test.Wilcox <- pairwiseWilcox(sce.filtered,
                              groups = sce.filtered$cell_type_manual,
                              block = sce.filtered$Sample,
                              BPPARAM = bpp
                              )

test.Wilcox$statistics[[2]] %>% 
  as.data.frame() %>% 
  rownames_to_column("genesymbol") %>% 
  filter(genesymbol == "CCND1")
```

# Save

```{r}
# save cell type annotation to merge with non-B cells later
df.manual.cell_types <- colData(sce.filtered) %>% 
  as.data.frame() %>% 
  dplyr::select(cell_type_manual) %>% 
  rownames_to_column("barcode_sample")
head(df.manual.cell_types)
```

```{r}
iteration_output = "3_Bcells"
storage.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_output)
dir.create(storage.dir, showWarnings = F, recursive = T)
saveRDS(sce.filtered,
        file = file.path(storage.dir, "sce.rds")
        )
saveRDS(df.manual.cell_types,
        file = file.path(storage.dir, "df.manual.cell_types.rds")
        )
write.xlsx(samples.df,
           file.path(storage.dir, "cell_metadata.xlsx")
           )
```

```{r}
sessionInfo()
```
