==> legacy/analysis/scripts/sc-processing/4_labelling.Rmd <==
---
title: "cycle 4 - Labeling all cells"
output: 
  html_notebook:
    toc: true
    toc_float: true
    code_folding: show
    theme: united
---

```{r setup}
library(tidyverse)
library(openxlsx)

# SC methods
library(scran)
library(scater)
library(bluster)

# Annotations
library(SingleR)
library(Seurat)
library(tricycle)

# Variable with random seed value used throughout the analysis for reproducibility
random_seed_value = 17

# Parallelization
library(BiocParallel)
bpp <- MulticoreParam()

# Path to project folder
wd <- here::here()
fig.outputs <- file.path(wd, "scRNAseq/Figures")
source(file.path(wd, "scRNAseq/analysis/scripts/colors.R"))

iteration_input = "2_cell_types"
sce.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_input)
```

```{r}
patients <- c("P009", "P022", "P027", "P087", "P069")
timepoints <- c("DG", "REL")
compartments <- c("PBMC", "BM", "LN", "TON", "GUT", "ASC")

samples.df <- read.xlsx(file.path(sce.dir, "cell_metadata.xlsx")) %>% 
  mutate(
    patient = factor(patient, levels = patients),
    timepoint = factor(timepoint, levels = timepoints),
    compartment = factor(compartment, levels = compartments),
    sample_norep = str_remove(sample, "_rep[0-9]"), .after = "sample"
  )
samples.df
```

# Load total cell dataset

```{r}
sce <- readRDS(file = file.path(sce.dir, "sce.rds"))
sce$sample_norep <- str_remove(sce$Sample, "_rep[0-9]")
sce$sample_norep <- factor(sce$sample_norep,
                           levels = unique(str_remove(samples.df$sample, "_rep[0-9]")))
sce$sample_norep_plot <- str_replace_all(sce$sample_norep, "_", " ")
sce$sample_norep_plot <- factor(sce$sample_norep_plot, levels = str_replace_all(unique(samples.df$sample_norep), "_", " "))
sce
```

## Load and assign B cell labels

```{r}
df.labels.Bcell <-
  readRDS(file = file.path(
    wd,
    "scRNAseq/analysis/outputs",
    "3_Bcells",
    "df.manual.cell_types.rds"
  ))
head(df.labels.Bcell)
```

```{r}
df.labels.full <- colData(sce) %>% 
  as.data.frame() %>% 
  dplyr::select("main_cell_type_manual" = cell_type_manual) %>% 
  rownames_to_column("barcode_sample") %>% 
  left_join(., df.labels.Bcell) %>% 
  mutate(
    main_cell_type_manual = case_when(
      is.na(cell_type_manual) ~ main_cell_type_manual,
      TRUE ~ cell_type_manual
    ),
    cell_type_manual = NULL
  )

df.labels.full
```

# Total dataset: Plot figures for presentations

UMAP overcorrects normal B-cells and P069 indolent cells

```{r fig.height=4.5, fig.width=12, message=FALSE}
ggpubr::ggarrange(
  plotReducedDim(sce,
                 dimred = "TSNE",
                 color_by = "sample_norep_plot") +
    scale_color_manual(values = get_palettes("tableau20")) +
    guides(color = guide_legend(override.aes = list(size = 4))) +
    labs(color = "Sample") +
    cowplot::theme_half_open(),
  plotReducedDim(sce,
                 dimred = "TSNE",
                 color_by = "cell_type_manual") +
    scale_color_manual(
      values = c(
        "B cell" = "#006BA4",
        "MCL" = "#D82526",
        "Monocyte" = "#32A251",
        "T cell" = "#FF7F0F"
      )
    ) +
    guides(color = guide_legend(override.aes = list(size = 4))) +
    cowplot::theme_half_open(),
  widths = c(1.05, 1)
)
```

```{r fig.height=4.5, fig.width=12, message=FALSE}
ggpubr::ggarrange(
  plotReducedDim(sce,
                 dimred = "TSNE",
                 color_by = "CCND1") +
    cowplot::theme_half_open(),
  
  plotReducedDim(sce,
                 dimred = "TSNE",
                 color_by = "cell_type_manual") +
    scale_color_manual(
      values = c(
        "B cell" = "#006BA4",
        "MCL" = "#D82526",
        "Monocyte" = "#32A251",
        "T cell" = "#FF7F0F"
      )
    ) +
    guides(color = guide_legend(override.aes = list(size = 4))) +
    cowplot::theme_half_open(),
  
  widths = c(0.95, 1)
)
```

Confirmation that the normal cells occupy same region on the tSNE

```{r, fig.height=6, fig.width=15}
plotReducedDim(
  sce,
  dimred = "TSNE",
  color_by = "cell_type_manual",
  other_fields = c("patient", "timepoint")
) +
  theme_bw() +
  facet_grid(timepoint ~ patient)
```

```{r, fig.width=6, fig.height=5}
plotReducedDim(
  sce, 
  dimred = "TSNE",
  color_by = "tricyclePhase",
  text_by = "cell_type_manual"
)
```

```{r, fig.width=10, fig.height=4}
plotReducedDim(
  sce, 
  dimred = "TSNE_corrected_SamplePhase",
  color_by = "cell_type_manual",
  text_by = "cell_type_manual"
) +
  plotReducedDim(
  sce, 
  dimred = "TSNE_corrected_SamplePhase",
  color_by = "tricyclePhase",
  text_by = "cell_type_manual"
)
```

# Count the cells

```{r}
samples.df <- colData(sce) %>%
  as.data.frame() %>%
  dplyr::select(Sample, cell_type_manual) %>%
  remove_rownames() %>%
  mutate(
    cell_type_manual = case_when(
      cell_type_manual == "B cell" ~ "N_cell_B",
      cell_type_manual == "MCL" ~ "N_cell_MCL",
      cell_type_manual == "T cell" ~ "N_cell_T",
      cell_type_manual == "Monocyte" ~ "N_cell_Monocytes"
    )
  ) %>%
  group_by(Sample, cell_type_manual) %>%
  summarise(n = n(), .groups = "drop") %>%
  pivot_wider(names_from = "cell_type_manual",
              values_from = "n",
              values_fill = 0) %>%
  left_join(samples.df, ., by = c("sample" = "Sample"))
samples.df
```

# Save labelled dataset

```{r}
iteration_output = "4_all_cells_labbeled"
storage.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_output)
dir.create(storage.dir, showWarnings = F, recursive = T)
saveRDS(sce,
        file = file.path(storage.dir, "sce.rds")
        )
write.xlsx(samples.df,
           file.path(storage.dir, "cell_metadata.xlsx")
           )
```

```{r}
sessionInfo()
```
