==> legacy/analysis/scripts/sc-processing/5_MCLcells.Rmd <==
---
title: "cycle 4 - MCL cells"
output: 
  html_notebook:
    toc: true
    toc_float: true
    code_folding: show
    theme: united
---

```{r setup}
library(tidyverse)
library(cowplot)
library(openxlsx)

# SC methods
library(scran)
library(scater)
library(bluster)
library(Seurat)

# Annotations
library(SingleR)
library(tricycle)

# Variable with random seed value used throughout the analysis for reproducibility
random_seed_value = 17

# Parallelization
library(BiocParallel)
bpp <- MulticoreParam()

# Path to project folder
wd <- here::here()
source(file.path(wd, "scRNAseq/analysis/scripts/colors.R"))

iteration_input = "4_all_cells_labbeled"
sce.dir <- file.path(wd, "scRNAseq/analysis/outputs", iteration_input)
```

```{r}
patients <- c("P009", "P022", "P027", "P087", "P069")
timepoints <- c("DG", "REL")
compartments <- c("PBMC", "BM", "LN", "TON", "GUT", "ASC")

samples.df <- read.xlsx(file.path(sce.dir, "cell_metadata.xlsx")) %>% 
  mutate(
    patient = factor(patient, levels = patients),
    timepoint = factor(timepoint, levels = timepoints),
    compartment = factor(compartment, levels = compartments)
  )
samples.df
```

```{r}
samples.df %>% 
  group_by(timepoint) %>% 
  summarise(cells = sum(N_cells_QCfiltered))
```

```{r}
samples.df %>% 
  summarise(cells = sum(N_cells_QCfiltered))
```


```{r}
sce <- readRDS(file = file.path(sce.dir, "sce.rds"))
sce
```

```{r, fig.width=9, fig.height=4}
chosen_resolution = "clusters_louvain_K30"

plotReducedDim(sce,
               dimred = "TSNE",
               color_by = "cell_type_manual",
               text_by = chosen_resolution) +
  plotReducedDim(sce,
                 dimred = "TSNE",
                 color_by = chosen_resolution,
                 text_by = chosen_resolution) +
  theme(legend.position = "none")
```

# Filtering

```{r}
sce.filtered <- sce[, sce$cell_type_manual == "MCL"]
sce.filtered
```

# Processing

## Normalization

```{r warning=FALSE, fig.height=5, fig.width=6}
set.seed(random_seed_value)
clusters_igraph <- quickCluster(sce.filtered,
                                method = "igraph",
                                BPPARAM = bpp)
sce.filtered$clusters_igraph <- clusters_igraph

plotReducedDim(sce.filtered,
               dimred = "PCA",
               color_by = "clusters_igraph")
```

```{r warning=FALSE}
set.seed(random_seed_value)
sce.filtered <- computeSumFactors(sce.filtered,
                                cluster = sce.filtered$clusters_igraph,
                                BPPARAM = bpp)
sce.filtered <- logNormCounts(sce.filtered,
                            size.factors = sce.filtered$sizeFactor,
                            BPPARAM = bpp)
sce.filtered
```

## HVG detection

```{r warning=FALSE}
# Feature selection
set.seed(random_seed_value)
variance_decomposition <- modelGeneVar(sce.filtered, BPPARAM = bpp)
top_HVG <- getTopHVGs(variance_decomposition,
                      fdr.threshold = 0.5)
length(top_HVG)
```

```{r fig.height=6, fig.width=10, warning=FALSE}
# I am removing IGLC2 and IGKC as their variation is given by the patient of origin and not cell biology
top_HVG <- top_HVG[!top_HVG %in% c("IGLC2", "IGKC")]

variance_df <- as.data.frame(variance_decomposition) %>%
  rownames_to_column("Symbol") %>%
  left_join(., as.data.frame(rowData(sce.filtered)), by = "Symbol") %>%
  mutate(HVG = case_when(Symbol %in% top_HVG ~ TRUE, TRUE ~ FALSE))

ggplot(variance_df,
       aes(x = mean, y = bio, color = HVG)) +
  scale_color_manual(values = c("black", "red")) +
  ggrepel::geom_text_repel(
    data = dplyr::filter(variance_df, bio > 0.75),
    aes(label = Symbol), min.segment.length = 0,
    color = "gray30"
  ) +
  geom_point(shape = 1) +
  cowplot::theme_cowplot() +
  labs(
    title = "Variance decomposition plot",
    x = "Mean of log-expression",
    y = "Variance of log-expression",
    color = "Biological\nvariance"
  )
```

# Dimensionality reduction

## PCA

```{r}
# PCA
set.seed(random_seed_value)
sce.filtered <- denoisePCA(sce.filtered,
                           subset.row = top_HVG,
                           technical = variance_decomposition)
percent.var <-
  attr(reducedDims(sce.filtered)[["PCA"]], "percentVar")

# chose a number of PCA dimensions which explain more than 0.5% variance
chosen.PC.num <-
  percent.var[which(percent.var > 0.5)] %>% length()

plot(percent.var,
     log = "y",
     xlab = "PC",
     ylab = "Variance explained (%)")
abline(v = chosen.PC.num, col = "blue")
```

```{r, fig.width=10, fig.height=9}
(plotPCA(sce.filtered, color_by = "Sample") +
  plotPCA(sce.filtered, color_by = "sum")) /
(plotPCA(sce.filtered, color_by = "detected") +
plotPCA(sce.filtered, color_by = "CCND1"))
```

## TSNE

```{r, fig.width=8, fig.height=3}
set.seed(random_seed_value)
sce.filtered <- runTSNE(
  sce.filtered,
  BPPARAM = bpp
)

plotTSNE(sce.filtered, color_by = "Sample") +
  plotTSNE(sce.filtered, color_by = "tricyclePhase")
```

# Batch correction

## For sample and CC phase

```{r warning=FALSE}
sce.filtered$correct_for <- paste(sce.filtered$Sample,
                                  sce.filtered$tricyclePhase, sep = "_")

set.seed(random_seed_value)
sce.filtered <- batchelor::correctExperiments(
  sce.filtered,
  batch = sce.filtered$correct_for,
  PARAM = batchelor::FastMnnParam(BPPARAM = bpp)
)
sce.filtered
```

# Plotting

## tSNE

```{r, fig.width=6.5, fig.height=5}
set.seed(random_seed_value)
sce.filtered <-
  runTSNE(
    sce.filtered,
    dimred = "corrected",
    name = "TSNE_corrected_SamplePhase",
    BPPARAM = bpp
  )
plotReducedDim(sce.filtered,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "Sample")
```

```{r, fig.width=13, fig.height=5}
plotReducedDim(sce.filtered,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "compartment",
               other_fields = c("patient", "timepoint")
               ) +
  facet_grid(timepoint ~ patient) +
  theme_bw()
```

```{r, fig.width=10, fig.height=2.5}
plotReducedDim(sce.filtered,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "patient",
               other_fields = c("patient")
               ) +
  facet_wrap(~ patient, nrow = 1)
```

```{r, fig.width=6, fig.height=5}
plotReducedDim(sce.filtered,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "tricyclePhase")
```

```{r, fig.width=11, fig.height=5}
plotReducedDim(sce.filtered,
               dimred = "TSNE_corrected_SamplePhase",
               color_by = "compartment",
               other_fields = "timepoint") +
  facet_wrap(~ timepoint)
```

## Clustering

```{r}
DR.cor.clusters.plots <- list()
cluster.metrics.plots <- list()
for (KNN in c(30)) {
  set.seed(random_seed_value)
  nn.clust <- clusterCells(
    sce.filtered,
    use.dimred = "PCA",
    full = TRUE,
    BLUSPARAM = bluster::SNNGraphParam(
      k = KNN,
      type = "rank",
      cluster.fun = "louvain",
      BPPARAM = bpp
    )
  )
  
  sce.filtered[[paste0("clusters_louvain_K", KNN)]] <-
    nn.clust$clusters
  
  
  DR.cor.clusters.plots[[paste0("clusters_louvain_K", KNN)]] <-
    plotReducedDim(
      sce.filtered,
      dimred = "TSNE_corrected_SamplePhase",
      color_by = paste0("clusters_louvain_K", KNN),
      text_by = paste0("clusters_louvain_K", KNN)
    ) +
    labs(title = paste0("kNN = ", KNN)) +
    plotReducedDim(
      sce.filtered,
      dimred = "UMAP_corrected_SamplePhase",
      color_by = paste0("clusters_louvain_K", KNN),
      text_by = paste0("clusters_louvain_K", KNN)
    )
  
  clusters <- nn.clust$clusters
  
  colors.louvain <-
    get_palettes("tableau20")[1:length(levels(clusters))]
  names(colors.louvain) <- levels(clusters)
  
  sil.approx <-
    bluster::approxSilhouette(reducedDim(sce.filtered, "PCA"),
                              clusters = clusters) %>%
    as.data.frame() %>%
    mutate(closest = factor(ifelse(width > 0, clusters, other)),
           cluster = clusters)
  
  purity <-
    bluster::neighborPurity(reducedDim(sce.filtered, "PCA"), clusters) %>%
    as.data.frame() %>%
    mutate(maximum = factor(maximum),
           cluster = clusters)
  
  sf.by.clust <- split(sizeFactors(sce.filtered), clusters) %>%
    vapply(., mean, 0) %>%
    data.frame("size.factors.mean" = .) %>%
    rownames_to_column("Cluster")
  
  rmsd <-
    bluster::clusterRMSD(reducedDim(sce.filtered, "PCA"), clusters) %>%
    data.frame("RMSD" = .) %>%
    rownames_to_column("Cluster") %>%
    left_join(., sf.by.clust, by = "Cluster")
  
  
  cluster.metrics.plots[[paste0("clusters_louvain_K", KNN)]] <-
    ggpubr::ggarrange(
      plotReducedDim(
        sce.filtered[, sce.filtered$patient != "P069"],
        dimred = "TSNE_corrected_SamplePhase",
        color_by = paste0("clusters_louvain_K", KNN),
        point_size = 0.5,
        other_fields = c("patient", "timepoint")
      ) +
        facet_grid(timepoint ~ patient) +
        theme_bw() +
        xlim(-35, 40) + ylim(-30, 36) +
        scale_color_manual(values = colors.louvain) +
        labs(color = "Cluster") +
        guides(color = guide_legend(nrow = 1, byrow = TRUE)),
      
      ggpubr::ggarrange(
        plotReducedDim(
          sce.filtered[, sce.filtered$patient == "P069"],
          dimred = "TSNE_corrected_SamplePhase",
          color_by = paste0("clusters_louvain_K", KNN),
          point_size = 0.5,
          other_fields = c("compartment", "timepoint")
        ) +
          facet_grid(timepoint ~ compartment) +
          theme_bw() +
          theme(legend.position = "none") +
          xlim(-35, 40) + ylim(-30, 36) +
          scale_color_manual(values = colors.louvain) +
          labs(color = "Cluster"),
        NULL,
        widths = c(3, 0.9)
      ),
      ggplot(sil.approx, aes(
        x = cluster, y = width, colour = closest
      )) +
        ggbeeswarm::geom_quasirandom(method = "smiley") +
        scale_color_manual(values = colors.louvain) +
        theme_bw(),
      
      ggplot(purity, aes(
        x = cluster, y = purity, colour = maximum
      )) +
        ggbeeswarm::geom_quasirandom(method = "smiley") +
        scale_color_manual(values = colors.louvain) +
        theme_bw(),
      
      ggplot(rmsd, aes(
        x = Cluster, y = RMSD, fill = Cluster
      )) +
        geom_col(color = "gray10") +
        scale_fill_manual(values = colors.louvain) +
        theme_bw(),
      
      ggplot(rmsd, aes(
        x = RMSD, y = size.factors.mean, fill = Cluster
      )) +
        geom_point(size = 3, shape = 21) +
        scale_fill_manual(values = colors.louvain) +
        theme_bw() + xlim(0, max(rmsd$RMSD)),
      
      ncol = 2,
      nrow = 3,
      common.legend = T,
      labels = c(
        paste("K=", KNN),
        "P069",
        "Silhouettes",
        "Purity",
        "Root-mean-squared-deviation",
        "RMSD vs avg. size factor"
      ),
      label.y = 1.05,
      font.label = list(
        size = 12,
        color = "black",
        face = "bold"
      )
    )
}
```


```{r, fig.width=10, fig.height=4}
DR.cor.clusters.plots
```

```{r fig.height=10, fig.width=13, message=FALSE, warning=FALSE}
cluster.metrics.plots
```

## DE testing

```{r}
MCL.Wilcox <- pairwiseWilcox(sce.filtered, 
                             groups = sce.filtered$timepoint,
                             lfc = 1,
                             BPPARAM = bpp
                             )
REL.Wilcox.df <- MCL.Wilcox$statistics[[2]] %>% 
  as.data.frame() %>% 
  rownames_to_column("genesymbol") %>% 
  filter(p.value < 0.05 | FDR < 0.25) %>% 
  arrange(p.value)
REL.Wilcox.df
```

```{r}
plots <- list()
for(feature in REL.Wilcox.df$genesymbol) {
  plots[[feature]] <- plotExpression(
    sce.filtered,
    features = feature,
    x = "timepoint",
    color = "timepoint",
    other_fields = "patient"
  ) +
    facet_wrap( ~ patient) +
    ggtitle(feature)
}
plots
```

```{r}
sce.aggr <- sce.filtered[, sce.filtered$patient != "P069"]

aggr.MCL.Wilcox <- pairwiseWilcox(sce.aggr, 
                             groups = sce.aggr$timepoint,
                             lfc = 1,
                             BPPARAM = bpp
                             )
REL.aggr.Wilcox.df <- aggr.MCL.Wilcox$statistics[[2]] %>% 
  as.data.frame() %>% 
  rownames_to_column("genesymbol") %>% 
  filter(p.value < 0.01 | FDR < 0.01) %>% 
  arrange(p.value)
REL.aggr.Wilcox.df
```

```{r}
aggr.MCL.Wilcox$statistics[[2]] %>% 
  as.data.frame() %>% 
  rownames_to_column("genesymbol") %>% 
  filter(genesymbol == "MKI67")
```



```{r}
plots <- list()
for(feature in REL.aggr.Wilcox.df$genesymbol) {
  plots[[feature]] <- plotExpression(
    sce.aggr,
    features = feature,
    x = "timepoint",
    color = "timepoint",
    other_fields = "patient"
  ) +
    facet_wrap( ~ patient) +
    ggtitle(feature)
}
plots
```

# Save

```{r}
iteration_output = "5_MCL"
storage.dir <-
  file.path(wd, "scRNAseq/analysis/outputs", iteration_output)
dir.create(storage.dir, showWarnings = F, recursive = T)
saveRDS(sce.filtered,
        file = file.path(storage.dir, "sce.rds"))
write.xlsx(samples.df,
           file.path(storage.dir, "cell_metadata.xlsx"))
```

```{r}
sessionInfo()
```
